+++ CHANGELOG.md
@@ -5,9 +5,10 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-## 3.0.0-alpha.81 (2024-08-11)
 
-* (theimo1221) Fix typo
+## **WORK IN PROGRESS**
+
+* (theimo1221) Invert Velux Shutter level
 
 ## 3.0.0-alpha.80 (2024-08-11)